requirements.txt
Django==1.11.7
pytz==2017.3
dj-database-url==0.4.2
dj-static==0.0.6
django-extensions==1.9.1
django-filter==1.0.4
django-guardian==1.4.8
django-map-widgets==0.1.9
django-toolbelt==0.0.1
djangorestframework==3.6.3
djangorestframework-gis==0.11.2
funcsigs==1.0.2
gunicorn==19.7.1
Markdown==2.6.8
mock==2.0.0
pbr==3.1.1
psycopg2==2.7.1
six==1.10.0
static3==0.7.0
urllib3==1.21.1
whitenoise==3.3.0